refactor: LottoRetailer의 역할을 LottoSeller로 합쳐라
